+++ Serialization/Json/JsonDicomConverter.cs
@@ -21,15 +21,18 @@ namespace Dicom.Serialization
     public class JsonDicomConverter : JsonConverter
     {
         private readonly bool _writeTagsAsKeywords;
+        private readonly bool _autoValidate;
         private readonly static Encoding _jsonTextEncoding = Encoding.UTF8;
 
         /// <summary>
         /// Initialize the JsonDicomConverter.
         /// </summary>
         /// <param name="writeTagsAsKeywords">Whether to write the json keys as DICOM keywords instead of tags. This makes the json non-compliant to DICOM JSON.</param>
-        public JsonDicomConverter(bool writeTagsAsKeywords = false)
+        /// <param name="autoValidate">Whether the content of DicomItems shall be validated as soon as they are added to the DicomDataset. </param>
+        public JsonDicomConverter(bool writeTagsAsKeywords = false, bool autoValidate = true)
         {
             _writeTagsAsKeywords = writeTagsAsKeywords;
+            _autoValidate = autoValidate;
         }
 
         #region JsonConverter overrides
@@ -100,6 +103,10 @@ namespace Dicom.Serialization
         private DicomDataset ReadJsonDataset(JToken obj)
         {
             var dataset = new DicomDataset();
+            if (!_autoValidate)
+            {
+                dataset = dataset.NotValidated();
+            }
             if (obj.Type == JTokenType.Null) { return null; }
             if (!(obj is JObject itemObject)) { throw new JsonReaderException("Malformed DICOM json"); }
 
@@ -489,7 +496,7 @@ namespace Dicom.Serialization
                 DicomValidation.ValidateDS(val);
                 return true;
             }
-            catch(DicomValidationException)
+            catch (DicomValidationException)
             {
                 return false;
             }
@@ -504,6 +511,9 @@ namespace Dicom.Serialization
         /// <returns>A json number equivalent to the supplied DS value</returns>
         private static string FixDecimalString(string val)
         {
+            // trim invalid padded character
+            val = val.Trim().TrimEnd('\0');
+
             if (IsValidJsonNumber(val))
             {
                 return val;
@@ -762,7 +772,7 @@ namespace Dicom.Serialization
             var childValues = new List<T>();
             foreach (var item in tokens)
             {
-                if (!(item.Type == JTokenType.String  && item.Value<string>() == "NaN") && !(item.Type  == JTokenType.Float || item.Type == JTokenType.Integer))
+                if (!(item.Type == JTokenType.String && item.Value<string>() == "NaN") && !(item.Type == JTokenType.Float || item.Type == JTokenType.Integer))
                 { throw new JsonReaderException("Malformed DICOM json"); }
                 childValues.Add((T)Convert.ChangeType(item.Value<object>(), typeof(T)));
             }