+++ ChangeLog.md
@@ -1,5 +1,11 @@
 #### v.4.0.7 (TBD)
 
+* Bug fix: DS VR type value ending with \0 fails to serialize in JsonDicomConverter (#1078)
+* Allow disabling validation in JsonDicomConverter
+* Bug fix: IPv6 issue in DesktopNetworkStream
+* Optimize DicomTag.GetHashCode() (#1088)
+* Bug fix: Prevent special characters in association requests from crashing Fellow Oak DICOM (#1104)
+
 #### v.4.0.6 (8/6/2020)
 * Update to DICOM Standard 2020b.
 * Handle N-Event requests in DicomClient (required for synchronous storage commitment) (#1001)